fix issue with server
